bug fixed.

+++ src/lib/lz4.c
@@ -737,8 +737,8 @@ int LZ4_uncompress(const char* source,
 		if unlikely(op-ref<STEPSIZE)
 		{
 #if LZ4_ARCH64
-			unsigned int dec2table[]={0, 0, 0, -1, 0, 1, 2, 3};
-			unsigned int dec2 = dec2table[op-ref];
+			size_t dec2table[]={0, 0, 0, -1, 0, 1, 2, 3};
+			size_t dec2 = dec2table[op-ref];
 #else
 			const int dec2 = 0;
 #endif
